requirements.txt
numpy
opencv-python
tensorflow-gpu==1.13.1
tensorflow-serving-api==1.13.1
